first upload of project
BrowseAndDownload is a VB.net project.  See the readme file for details,
and also note that each screen of the program has a HELP button to
explain what it does. (Both the main folder and its subfolder have the
same name 'BrowseAndDownload', but the outer folder has the solution
file, and the subfolder has the VB files.